[BOSK-35] Moved slot.py to bosk/block
