+++ bosk/block/zoo/input_plugs/input_plugs.py
@@ -1,7 +1,7 @@
 from bosk.block import BaseBlock, BlockInputData, TransformOutputData
 from bosk.stages import Stages
 from bosk.block import BlockMeta
-from bosk.slot import InputSlotMeta, OutputSlotMeta
+from bosk.block.slot import InputSlotMeta, OutputSlotMeta
 from bosk.block.meta import make_simple_meta
