+++ bosk/executor/topological.py
@@ -11,7 +11,7 @@ from ..data import Data
 from .base import BaseExecutor
 from ..pipeline import BasePipeline
 from .descriptor import HandlingDescriptor
-from ..slot import BaseSlot, BlockInputSlot, BlockOutputSlot
+from ..block.slot import BaseSlot, BlockInputSlot, BlockOutputSlot
 from ..block import BaseBlock
 from .utility import get_connection_map
