+++ bosk/painter/graphviz.py
@@ -1,7 +1,7 @@
 from typing import Optional, Sequence
 from ..block.base import BaseBlock
 from ..pipeline.connection import Connection
-from ..slot import BlockInputSlot
+from ..block.slot import BlockInputSlot
 from .base import BasePainter
 from ..pipeline.base import BasePipeline
 from ..executor.base import BaseExecutor
